+++ src/lib.rs
@@ -10,6 +10,8 @@
 pub mod action;
 pub mod api;
 pub mod app;
+pub mod bee_log;
+pub mod bee_log_tailer;
 pub mod bee_supervisor;
 pub mod cli;
 pub mod components;